backend/requirements.txt
django
django-environ
djangorestframework
dj-rest-auth
djangorestframework-simplejwt
Pillow
django-extensions